LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/gap-expected.txt


PASS Can set 'column-gap' to CSS-wide keywords
FAIL Can set 'column-gap' to var() references assert_equals: expected 2 but got 1
PASS Can set 'column-gap' to the 'normal' keyword
FAIL Can set 'column-gap' to a length Invalid values
FAIL Can set 'column-gap' to a percent Invalid values
PASS Setting 'column-gap' to a time throws TypeError
PASS Setting 'column-gap' to an angle throws TypeError
PASS Setting 'column-gap' to a flexible length throws TypeError
FAIL Setting 'column-gap' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'column-gap' to a URL throws TypeError
PASS Setting 'column-gap' to a transform throws TypeError
PASS Can set 'row-gap' to CSS-wide keywords
FAIL Can set 'row-gap' to var() references assert_equals: expected 2 but got 1
PASS Can set 'row-gap' to the 'normal' keyword
FAIL Can set 'row-gap' to a length Invalid values
FAIL Can set 'row-gap' to a percent Invalid values
PASS Setting 'row-gap' to a time throws TypeError
PASS Setting 'row-gap' to an angle throws TypeError
PASS Setting 'row-gap' to a flexible length throws TypeError
FAIL Setting 'row-gap' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'row-gap' to a URL throws TypeError
PASS Setting 'row-gap' to a transform throws TypeError
